+++ cognee/modules/data/extraction/knowledge_graph/extract_content_graph.py
@@ -15,36 +15,3 @@ async def extract_content_graph(content: str, response_model: Type[BaseModel]):
     )
 
     return content_graph
-
-
-async def extract_content_edges(
-    content: str, potential_nodes: List[str], potential_relationships: List[str]
-) -> KnowledgeGraph:
-    """Creates a knowledge graph by identifying relationships between the provided potential nodes."""
-    llm_client = get_llm_client()
-
-    context = {
-        "potential_nodes": potential_nodes,
-        "potential_relationships": potential_relationships,
-    }
-
-    system_prompt = render_prompt("extract_graph_edges_prompt.txt", context)
-    graph = await llm_client.acreate_structured_output(
-        content, system_prompt, response_model=KnowledgeGraph
-    )
-
-    # Create a map of node names to ensure we don't duplicate nodes
-    node_map = {}
-    for node in graph.nodes:
-        if node.name not in node_map:
-            node_map[node.name] = node.id
-
-    # Filter edges to ensure they only connect existing nodes
-    valid_edges = []
-    for edge in graph.edges:
-        source_exists = any(node.id == edge.source_node_id for node in graph.nodes)
-        target_exists = any(node.id == edge.target_node_id for node in graph.nodes)
-        if source_exists and target_exists:
-            valid_edges.append(edge)
-
-    return KnowledgeGraph(nodes=list(graph.nodes), edges=valid_edges)
